Whoops, deadband is a double, not an integer

+++ src/main/java/com/team3316/kit/motors/DBugTalon.java
@@ -38,7 +38,7 @@ public class DBugTalon extends TalonSRX {
     this.configPeakOutputForward(+1, DBugTalon.kTimeout);
     this.configPeakOutputReverse(-1, DBugTalon.kTimeout);
     this.configNeutralDeadband(
-      (int) Config.getInstance().get(configLabel + ".neutralDeadband"),
+      (double) Config.getInstance().get(configLabel + ".neutralDeadband"),
       DBugTalon.kTimeout
     );
